+++ redGulp/extensions/plugins/redshop_pdf/tcpdf.js
@@ -11,6 +11,5 @@ var composer = require("gulp-composer");
 
 // Composer
 gulp.task("composer:" + baseTask, function (cb) {
-    composer({ cwd: extPath + "/helper", bin: "php ./composer.phar" });
-    cb();
+    composer({ cwd: extPath + "/helper", bin: "php ./composer.phar" }).on('end', cb);
 });
