--- redGulp/tasks/release/redshop.js
@@ -1,145 +1,0 @@
-var gulp = require("gulp");
-var path = require("path");
-var gutil = require('gulp-util');
-var zip = require("gulp-zip");
-var fs = require("fs");
-var del = require('del');
-var sass = require('gulp-sass');
-var rename = require('gulp-rename');
-var concat = require('gulp-concat');
-var uglify = require('gulp-uglify');
-// Get console args
-var argv = require("yargs").argv;
-// XML parser
-var xml2js = require("xml2js");
-var parser = new xml2js.Parser();
-require('./../../extensions/components/redshop');
-require('./../../extensions/components/redshop.assets');
-var config = require('./../../../gulp-config.json');
-
-
-/// Define component tasks
-var componentName = 'com_redshop';
-
-/// Define paths of source and destination
-var extPath = '.';
-var mediaPath = extPath + '/media/' + componentName;
-var assetsPath = extPath + '/src/assets/' + componentName;
-
-function getIncludedExtensions() {
-    var included = [];
-    var includedPlugins = [
-        "./plugins/content/rs_menuitem_sh404urls/**",
-        "./plugins/system/redshop/**",
-        "./plugins/system/redgoogleanalytics/**",
-        "./plugins/redshop_payment/rs_payment_banktransfer/**",
-        "./plugins/redshop_payment/rs_payment_paypal/**",
-        "./plugins/redshop_product/sh404urls/**",
-        "./plugins/finder/redshop/**",
-        "./plugins/redshop_alert/alert/**",
-        "./plugins/redshop_shipping/default_shipping/**",
-        "./plugins/sh404sefextplugins/sh404sefextplugincom_redshop/**",
-        "./plugins/redshop_pdf/tcpdf/**",
-        "./plugins/redshop_export/attribute/**",
-        "./plugins/redshop_export/category/**",
-        "./plugins/redshop_export/field/**",
-        "./plugins/redshop_export/manufacturer/**",
-        "./plugins/redshop_export/newsletter_subscriber/**",
-        "./plugins/redshop_export/product/**",
-        "./plugins/redshop_export/product_stockroom_data/**",
-        "./plugins/redshop_export/related_product/**",
-        "./plugins/redshop_export/shipping_address/**",
-        "./plugins/redshop_export/shopper_group_attribute_price/**",
-        "./plugins/redshop_export/shopper_group_product_price/**",
-        "./plugins/redshop_export/user/**",
-        "./plugins/redshop_export/order/**",
-        "./plugins/redshop_import/attribute/**",
-        "./plugins/redshop_import/category/**",
-        "./plugins/redshop_import/field/**",
-        "./plugins/redshop_import/manufacturer/**",
-        "./plugins/redshop_import/newsletter_subscriber/**",
-        "./plugins/redshop_import/product/**",
-        "./plugins/redshop_import/product_stockroom_data/**",
-        "./plugins/redshop_import/shipping_address/**",
-        "./plugins/redshop_import/shopper_group_product_price/**",
-        "./plugins/redshop_import/shopper_group_attribute_price/**",
-        "./plugins/redshop_import/user/**",
-        "./plugins/redshop_import/related_product/**",
-        "./plugins/quickicon/redshop/**",
-        "./plugins/search/redshop_categories/**",
-        "./plugins/search/redshop_products/**",
-        "./plugins/twig/cache/**",
-        "./plugins/twig/debug/**",
-        "./plugins/twig/japp/**",
-        "./plugins/twig/jarray/**",
-        "./plugins/twig/jdoc/**",
-        "./plugins/twig/jhtml/**",
-        "./plugins/twig/jlang/**",
-        "./plugins/twig/jlayout/**",
-        "./plugins/twig/jmodule/**",
-        "./plugins/twig/jposition/**",
-        "./plugins/twig/jprofiler/**",
-        "./plugins/twig/jregistry/**",
-        "./plugins/twig/jroute/**",
-        "./plugins/twig/jsession/**",
-        "./plugins/twig/jtext/**",
-        "./plugins/twig/juri/**",
-        "./plugins/twig/juser/**",
-        "./plugins/twig/unseriallize/**",
-        "./plugins/twig/redshop/**",
-    ];
-    var excluded = [
-        '!./plugins/redshop_pdf/tcpdf/helper/vendor/tecnickcom/tcpdf/fonts/!(courier*.php|helvetica*.php|symbol*.php|times*.php|uni2cid_a*.php|zapfdingbats*.php)',
-        '!./plugins/redshop_pdf/tcpdf/helper/vendor/tecnickcom/tcpdf/fonts/*/*'
-    ];
-    var includedModules = [
-        "./modules/site/mod_redshop_cart/**",
-        "./modules/site/mod_redshop_search/**"
-    ];
-
-    return included.concat(includedPlugins, includedModules, excluded);
-}
-
-// Composer
-gulp.task('clean:libraries.redshop:composer.lock', function (cb) {
-    del(extPath + '/composer.lock', { force: true });
-    cb();
-})
-
-gulp.task("composer:libraries.redshop", gulp.series('clean:libraries.redshop:composer.lock'), function (cb) {
-    executeComposer(extPath);
-    cb();
-});
-
-/// Minified and deploy from Src to Media.
-gulp.task('scripts:components.redshop', function () {
-    return gulp.src([
-        assetsPath + '/js/*.js',
-        assetsPath + '/js/**/*.js'
-    ])
-        .pipe(gulp.dest(mediaPath + '/js'))
-        .pipe(uglify())
-        .on('error', function (err) { gutil.log(gutil.colors.red('[Error]'), err.toString()); })
-        .pipe(rename(function (path) {
-            path.basename += '.min';
-        }))
-        .pipe(gulp.dest(mediaPath + '/js'));
-});
-
-/// Sass Compiler
-gulp.task('sass:components.redshop', function () {
-    return gulp.src([
-        assetsPath + "/scss/*.scss",
-        assetsPath + "/scss/**/*.scss"
-    ])
-        .pipe(sass())
-        .pipe(gulp.dest(mediaPath + "/css"))
-        .pipe(sass({
-            outputStyle: "compressed",
-            errLogToConsole: true
-        }))
-        .pipe(rename(function (path) {
-            path.basename += '.min';
-        }))
-        .pipe(gulp.dest(mediaPath + "/css"));
-});
